read_helper_desktop/requirements.txt
Flask==3.1.0
pywebview==5.4
pyinstaller==6.13.0
